requirements/common.txt
django
python-telegram-bot
python-dotenv